Refactored idEventArgs to not allocate memory dynamically at static initialization time. Simple store args in idStaticList instead of std::vector.
Note: my original intent was to remove dynamic allocations during static initialization (which is cleaner way of using memory).
But I see now that AI code creates and fills a lot of STL containers during static initialization time.
So it is not possible to achieve...
But I decided to commit this change anyway.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@7047 49c82d7f-2e2a-0410-a16f-ae8f201b507f